% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{creb1_sels}
\alias{creb1_sels}
\title{a selection of CREB1-targeted ChIP-seq 'signal p-value' reports in a GRangesList}
\format{GRangesList}
\usage{
creb1_sels
}
\description{
a selection of CREB1-targeted ChIP-seq 'signal p-value' reports in a GRangesList
}
\keyword{datasets}
