% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwapp.R
\name{genes_from_pathway}
\alias{genes_from_pathway}
\title{app to retrieve genes corresponding to one of a selection of pathways named in Gene Ontology}
\usage{
genes_from_pathway()
}
\description{
app to retrieve genes corresponding to one of a selection of pathways named in Gene Ontology
}
\note{
Uses org.Hs.eg.db and GO.db to obtain pathway names and gene mappings.
}
