% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gopw_selected}
\alias{gopw_selected}
\title{a selection of pathway-related terms in GO}
\format{character vector}
\usage{
gopw_selected
}
\description{
a selection of pathway-related terms in GO
}
\keyword{datasets}
