% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixpl.nnet.R
\name{plot.nnet}
\alias{plot.nnet}
\title{slightly modified program from beckmw gist (see note below)}
\usage{
plot.nnet(
  mod.in,
  nid = T,
  all.out = T,
  all.in = T,
  bias = T,
  wts.only = F,
  rel.rsc = 5,
  circle.cex = 5,
  node.labs = T,
  var.labs = T,
  x.lab = NULL,
  y.lab = NULL,
  line.stag = NULL,
  struct = NULL,
  cex.val = 1,
  alpha.val = 1,
  circle.col = "lightblue",
  pos.col = "black",
  neg.col = "grey",
  ...
)
}
\arguments{
\item{mod.in}{nnet instance}
}
\description{
slightly modified program from beckmw gist (see note below)
}
\note{
\url{https://tinyurl.com/tuyvbkk} is slightly broken, and this is a simple repair.
The original link is from \url{https://www.r-bloggers.com/visualizing-neural-networks-in-r-update/}.
}
