% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_mlr3.R
\name{se_to_filtered_task}
\alias{se_to_filtered_task}
\title{produce post-filter Tasks for mlr3}
\usage{
se_to_filtered_task(
  taskid,
  se,
  classvar,
  nfeat = 100,
  tmaker = mlr3::TaskClassif,
  filtfun = function(x) flt("variance")$calculate(x, nfeat = nfeat),
  ...
)
}
\arguments{
\item{taskid}{character(1) arbitrary label}

\item{se}{SummarizedExperiment instance}

\item{classvar}{character(1) optional name of colData element to use for classification}

\item{nfeat}{numeric(1) number of features to retain}

\item{tmaker}{a Task as defined in mlr3}

\item{filtfun}{a function that computes a Filter score in the mlr3filters framework, defaults
to the "variance" scoring method}

\item{\dots}{passed to `tmaker$new()` for the filtered task production}
}
\description{
produce post-filter Tasks for mlr3
}
