% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{skd}
\alias{skd}
\title{output of import_enc_bw with CREB1-targeted ChIP-seq 'signal p-value' reports}
\format{GRangesList}
\usage{
skd
}
\description{
output of import_enc_bw with CREB1-targeted ChIP-seq 'signal p-value' reports
}
\keyword{datasets}
