% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uu}
\alias{uu}
\title{output of import_enc_bw with FOXA1-targeted ChIP-seq 'signal p-value' reports}
\format{GRangesList}
\usage{
uu
}
\description{
output of import_enc_bw with FOXA1-targeted ChIP-seq 'signal p-value' reports
}
\keyword{datasets}
